Extract ConnectionHandler to a separate file and rename to handler::TcpConnection
