+++ src/net/mod.rs
@@ -1,5 +1,6 @@
-pub use self::server::Server;
-pub use self::thread::{Job, Message, Pool};
+pub use self::server::IPAService;
+pub use self::thread::{Message, Pool};
 
+mod handler;
 mod server;
 mod thread;
