+++ src/net/thread.rs
@@ -14,7 +14,7 @@ struct Worker {
     thread: Option<thread::JoinHandle<()>>,
 }
 
-pub type Job = Box<dyn FnOnce() + Send + 'static>;
+type Job = Box<dyn FnOnce() + Send + 'static>;
 
 pub enum Message {
     NewJob(Job),
